+++ packages/zowe-explorer/__tests__/__unit__/abstract/TreeProvider.unit.test.ts
@@ -65,6 +65,8 @@ async function createGlobalMocks() {
         mockProfilesCache: new ProfilesCache(imperative.Logger.getAppLogger()),
     };
 
+    Object.defineProperty(globals, "LOG", { value: jest.fn(), configurable: true });
+    Object.defineProperty(globals.LOG, "error", { value: jest.fn(), configurable: true });
     Object.defineProperty(globalMocks.mockProfilesCache, "getProfileInfo", {
         value: jest.fn(() => {
             return { value: globalMocks.mockProfileInfo, configurable: true };