+++ packages/zowe-explorer/src/uss/ZoweUSSNode.ts
@@ -377,6 +377,7 @@ export class ZoweUSSNode extends ZoweTreeNode implements IZoweUSSTreeNode {
                 }
             } catch (err) {}
         } catch (err) {
+            globals.LOG.error(err);
             vscode.window.showErrorMessage(
                 localize("deleteUSSNode.error.node", "Unable to delete node: ") + err.message
             );
@@ -584,6 +585,7 @@ export class ZoweUSSNode extends ZoweTreeNode implements IZoweUSSTreeNode {
             }
         } catch (err) {
             if (err.message.includes(localize("refreshUSS.error.notFound", "not found"))) {
+                globals.LOG.warn(err);
                 vscode.window.showInformationMessage(
                     localize("refreshUSS.file1", "Unable to find file: ") +
                         label +
@@ -603,6 +605,7 @@ export class ZoweUSSNode extends ZoweTreeNode implements IZoweUSSTreeNode {
             try {
                 document = await vscode.workspace.openTextDocument(documentPath);
             } catch (err) {
+                globals.LOG.warn(err);
                 openingTextFailed = true;
             }