+++ packages/zowe-explorer/src/utils/ProfilesUtils.ts
@@ -45,6 +45,7 @@ export async function errorHandling(errorDetails: any, label?: string, moreInfo?
         "errorHandling.invalid.token",
         "Your connection is no longer active. Please log in to an authentication service to restore the connection."
     );
+    [errorDetails, errorDetails.mDetails, label, moreInfo].filter(Boolean).forEach(globals.LOG.error);
 
     if (errorDetails.mDetails !== undefined) {
         httpErrCode = errorDetails.mDetails.errorCode;
@@ -153,6 +154,7 @@ export const syncSessionNode =
         try {
             profile = Profiles.getInstance().loadNamedProfile(profileName, profileType);
         } catch (e) {
+            globals.LOG.warn(e);
             return;
         }
         sessionNode.setProfileToChoice(profile);
@@ -174,10 +176,11 @@ export interface IFilterItem {
     description?: string;
     show?: boolean;
     icon?: string;
+    menuType?: globals.JobPickerTypes;
 }
 
 export class FilterItem implements vscode.QuickPickItem {
-    constructor(private filterItem: IFilterItem) {}
+    constructor(public filterItem: IFilterItem) {}
     get label(): string {
         const icon = this.filterItem.icon ? this.filterItem.icon + " " : null;
         return (icon ?? "") + this.filterItem.text;